v4.1.3 - New poof functions. New config values.
Default "poof"-values are modified. More sidepoof and less height poof
now. Added option which increases the height and sidepoof if you jump on
mushrooms while sprinting. This is on by default. Improved code. You
will now be side-pood'd in the angle you are facing, and not forced to
N,S,E,W,SE,NW,SW or NE.